# Use a C abi interface between ykrt and the rest of yorick