--- a/ykrt/Cargo.toml
+++ b/ykrt/Cargo.toml
@@ -6,5 +6,4 @@ edition = "2018"
 license = "Apache-2.0 OR MIT"
 
 [dependencies]
-yktrace = { path = "../yktrace" }
-ykcompile = { path = "../ykcompile" }
+ykrt_internal = { path = "../ykrt_internal" }