--- a/ykrt/src/internal.rs
+++ b/ykrt/src/internal.rs
@@ -1,4 +1,6 @@
-use std::{ffi::c_void, marker::PhantomData};
+use std::ffi::{CString, c_void};
+use std::marker::PhantomData;
+use std::os::raw::c_char;
 
 // Force ykrt_internal to be linked
 #[allow(unused_imports)]
@@ -6,8 +8,8 @@ use ykrt_internal;
 
 extern "C" {
     fn __ykrt_start_tracing(tracing_kind: u8) -> *mut c_void;
-    fn __ykrt_stop_tracing(tracer: *mut c_void) -> *mut c_void;
-    fn __ykrt_compile_trace(sir_trace: *mut c_void) -> *mut c_void;
+    fn __ykrt_stop_tracing(tracer: *mut c_void, error_msg: *mut *mut c_char) -> *mut c_void;
+    fn __ykrt_compile_trace(sir_trace: *mut c_void, error_msg: *mut *mut c_char) -> *mut c_void;
     fn __ykrt_compiled_trace_get_ptr(compiled_trace: *const c_void) -> *const c_void;
     fn __ykrt_compiled_trace_dealloc(compiled_trace: *mut c_void);
 }
@@ -25,12 +27,19 @@ pub enum TracingKind {
 pub(crate) struct ThreadTracer(*mut c_void);
 
 pub(crate) fn start_tracing(tracing_kind: TracingKind) -> ThreadTracer {
-    unsafe { ThreadTracer(__ykrt_start_tracing(tracing_kind as u8)) }
+    let tracer = unsafe { __ykrt_start_tracing(tracing_kind as u8) };
+    assert!(!tracer.is_null());
+    ThreadTracer(tracer)
 }
 
 impl ThreadTracer {
-    pub(crate) fn stop_tracing(self) -> SirTrace {
-        unsafe { SirTrace(__ykrt_stop_tracing(self.0)) }
+    pub(crate) fn stop_tracing(self) -> Result<SirTrace, CString> {
+        let mut err_msg = std::ptr::null_mut();
+        let sir_trace = unsafe { __ykrt_stop_tracing(self.0, &mut err_msg) };
+        if sir_trace.is_null() {
+            return Err(unsafe { CString::from_raw(err_msg) });
+        }
+        Ok(SirTrace(sir_trace))
     }
 }
 
@@ -47,13 +56,16 @@ pub(crate) struct CompiledTrace<I> {
 unsafe impl<I> Send for CompiledTrace<I> {}
 unsafe impl<I> Sync for CompiledTrace<I> {}
 
-pub(crate) fn compile_trace<T>(sir_trace: SirTrace) -> CompiledTrace<T> {
-    unsafe {
-        CompiledTrace {
-            compiled: __ykrt_compile_trace(sir_trace.0),
-            _marker: PhantomData,
-        }
+pub(crate) fn compile_trace<T>(sir_trace: SirTrace) -> Result<CompiledTrace<T>, CString> {
+    let mut err_msg = std::ptr::null_mut();
+    let compiled = unsafe { __ykrt_compile_trace(sir_trace.0, &mut err_msg) };
+    if compiled.is_null() {
+        return Err(unsafe { CString::from_raw(err_msg) });
     }
+    Ok(CompiledTrace {
+        compiled,
+        _marker: PhantomData,
+    })
 }
 
 impl<I> CompiledTrace<I> {