In the future this will make it possible to separately compile the internals of yorick with optimizations and without tracing support.