@bjorn3 2020-12-19
I am not quite sure yet how it should be compiled with a different compiler yet. I tried using a build script, but that will require recompiling everything on every change. In addition it will hide the output of the compilation process unless `-vv` is passed to cargo. I guess in the future it can become part of the rustc sysroot itself. For now maybe a script could be used to compile the internals first and then to pass the right link arguments when compiling the user of yorick?