@ltratt 2020-12-19
This looks interesting, and I understand the motivation, but maybe you can spell out any possible consequences too?