@bjorn3 2020-12-19
In the future it may be necessary to extend this interface, which could make it a bit harder to develop new features. I don't expect this to be necessary for the foreseeable future though. It also increases the amount of unsafe code.