@ltratt 2020-12-19
> In the future this will make it possible to separately compile the internals of yorick with optimizations and without tracing support.

Possibly stupid question: how does this make this feasible? I have a couple of guesses, but I'm not sure which might be correct (maybe none of them!).