@bjorn3 2020-12-19
Rust doesn't have a stable abi, so if you want to call code compiled by one version of rustc from code compiled by a different version of rustc, you must use the C abi or another stable abi.