@ltratt 2020-12-19
I see. The next question is whether putting the division line in ykrt is necessary and/or sufficient. I don't immediately know the answer to that, I must admit! @vext01 might?