@vext01 2020-12-19
So this is all so that we can compile most of the yorick with one set of optimisation flags, and the user's interpreter with another. Am I right?

I'm probably not going to get a chance to do a full review until Monday.