@vext01 2020-12-21
> I am not quite sure yet how it should be compiled with a different compiler yet.

A couple of thoughts on this.

Instead of invoking rustc directly, could the bits we want to compile separately be managed by a separate Cargo workspace? Then we would get incremental compilation, I think?

Also, we could use the same ykrustc compiler, just remove the `-C tracer=hw` flags.

Thanks for this PR, I'll review it in full now.