@vext01 2020-12-21
I think this is likely to be a (very) worthwhile change, but I'm a little hesitant to merge until a) we know we can get the build system details to work for us, and b) that we see a good speed up from optimising the JIT internals.

I wonder if we should flesh out the build system changes in this PR and then do some quick and dirty benchmarking with multitime?

@ptersilie do you have any thoughts?