@vext01 2020-12-21
I spoke to @ptersilie offline. He raised some other concerns (he's afk, but said I could echo here):

 * Would the new separated crate be statically linked? We don't want to carry around >1 artefact if we can help it.
 * Would this interfere with instantiating the blackhole interpreter? I (Edd), think this should be OK, as the interpreter would too be contained in side the "internal" crate, I think.
 * Does this change our development workflow much?

I hope that covers @ptersilie's concerns. Thanks!