@bjorn3 2020-12-21
> Would the new separated crate be statically linked? We don't want to carry around >1 artefact if we can help it.

Using `staticlib` as crate type will cause all libstd symbols to leak. If the libstd is not the same, this isn't a problem for most symbols as they will be mangled differently, but there are some symbols that aren't mangled and thus would still conflict. Using a `cdylib` would solve the symbol leaking issue, but would require the final executable to be shipped with a dynamic library.

On Linux it is possible to link the static library using `ld -r` to a single object file for which you can then hide all symbols except for the ones that need to be exported using objcopy. On macOS this results in a linker error when trying to link against the object file for some reason. This is a problem for cg_clif too as it needs it to compile inline asm using an external assembler and then link both the asm object and the regular object as the incremental system expects a single object file for each codegen unit.

> Would this interfere with instantiating the blackhole interpreter? I (Edd), think this should be OK, as the interpreter would too be contained in side the "internal" crate, I think.

Yes, the interpreter should also be inside the "internal" crate. I think it may be a good idea to make a sketch of the C api before this PR is merged.

> Does this change our development workflow much?

That is a good question. All tests that use tracing to get TIR need to be moved to ykrt or a crate that depends on ykrt. As the rest of the crates will no longer be built with tracing support. In addition the internal crate will somehow need to be built separately from ykrt. Finding a way to do this without hurting development is pretty important. For cg_clif I use a bunch of scripts to do everything from building cg_clif and building the sysroot to running tests. It works, but I don't know if you will like it as much as I do.