@vext01 2020-12-21
> Using staticlib as crate type will cause all libstd symbols to leak.

And I suppose we would encounter the same issue if we use a rlib?

I wonder if there is an alternative solution that doesn't require restructuring so much. Could we get ykrustc to recognise the `yktrace` crate (and friends) and optimise all of their items which are not marked `#[optimise(never)`? (It would probably require extending the `optimise` attribute, or adding our own attribute).