@bjorn3 2020-12-21
> Could we get ykrustc to recognise the yktrace crate (and friends) and optimise all of their items which are not marked #[optimise(never)?

It is not just yktrace, but all of it's dependencies, ykcompile and ykbh. Basically everything except for ykrt and the user crate/dependencies.