@vext01 2020-12-21
I see.

The staticlib problem is concerning. We would like this to run on other systems in the (probably) distant future, so we probably don't want to constrain ourselves too much if we can help it.