@vext01 2020-12-21
And does using a shared object solve these issues? The option to install into the compiler's sysroot could still be a path forward?