@ptersilie 2020-12-21
> Yes, the interpreter should also be inside the "internal" crate. I think it may be a good idea to make a sketch of the C api before this PR is merged.

I'm currently working on the blackholing stuff and the goal with the interpreter is to be instantiated by the JIT (upon guard failure) and then the interpreter instance is handed to the meta-tracer who runs it. Would this PR interfere with that? There is of course room to change the way the interpreter is instantiated and ran, though conceptually I dislike the idea of the compiler running the interpreter. But I guess that's a price I'm willing to pay to make the compiler/interpreter more performant.

I think sketching out the ABI would be a very good idea at this point, to see how this affects the rest of the system.