@bjorn3 2020-12-21
> And does using a shared object solve these issues? The option to install one into the compiler's sysroot could still be a path forward?

Yes, a cdylib works.

> I'm currently working on the blackholing stuff and the goal with the interpreter is to be instantiated by the JIT (upon guard failure) and then the interpreter instance is handed to the meta-tracer who runs it. Would this PR interfere with that? There is of course room to change the way the interpreter is instantiated and ran, though conceptually I dislike the idea of the compiler running the interpreter. But I guess that's a price I'm willing to pay to make the compiler/interpreter more performant.

The interpreter would have to export a C abi function anyway for the compiled trace to switch to it when a guard fails.

> I think sketching out the ABI would be a very good idea at this point, to see how this affects the rest of the system.

I think the first step would be to determine how to switch from the compiled trace to the blackhole interpreter.