@vext01 2021-01-12
Hi @bjorn3,

Your idea gives a pretty decent speed up. Enough for us to consider moving this forward.

Here are timings (over 30 runs) for the ykrt test suite on my branch (skipping two tests which crash trying to get a PT file descriptor):
```
Before:
1: env RUST_TEST_THREADS=1 target/debug/deps/ykrt-589c0db448b16d4d
            Mean        Std.Dev.    Min         Median      Max
real        2.648       0.061       2.526       2.661       2.766
user        24.545      0.248       23.971      24.612      25.047
sys         0.204       0.022       0.161       0.204       0.248

1: target/debug/deps/ykrt-589c0db448b16d4d
            Mean        Std.Dev.    Min         Median      Max
real        2.496       0.112       2.307       2.493       2.818
user        25.704      0.502       24.781      25.825      26.609
sys         0.227       0.056       0.092       0.226       0.401

After:
1: env RUST_TEST_THREADS=1 target/debug/deps/ykrt-b424f141bb012e29
            Mean        Std.Dev.    Min         Median      Max
real        1.672       0.148       1.557       1.613       2.104
user        22.127      0.705       21.383      21.959      24.410
sys         0.053       0.012       0.036       0.048       0.088

1: target/debug/deps/ykrt-b424f141bb012e29
            Mean        Std.Dev.    Min         Median      Max
real        1.613       0.027       1.565       1.612       1.674
user        22.359      0.103       22.136      22.366      22.539
sys         0.059       0.015       0.036       0.056       0.100
```

Having worked on this, I have fresh questions.

> The problem is not that you can't mix unoptimized and optimized code, but that you must not link two separately compiled libstd's into a single executable or dylib and must not try to interpret a type from one of the compilations as one from the other compilation as would be necessary when using libstd types in the ffi function signatures.

I'm still really keen to use the rust ABI if we can make it at all possible.

So you are saying that we have one libstd for the optimised portion inside the cydlib and another copy of libstd for the unoptimised portion and we can't leak types from one stdlib across the optimisation threshold due to potential incompatibilities introduced by `-C tracer=hw`.

Currently the std lib in the ykrustc sysroot will only ever be compiled *without* `-C tracer=hw`. We don't do anything special for the stdlib right now, as we are unable to trace it anyway. So both the optimised and unoptimised portions will use the same sole stdlib but each will have their own copy. Does that mean that we can relax the constraints and in fact use the rust ABI instead of the C ABI?

There is of course a question of whether we might be able to build an unoptimised stdlib at some time in the future, thus allowing tracing the stdlib. At that point we may link a different stdlib in the optimised and unoptimised portions. At that point, we clearly shouldn't use the rust ABI.

So in fact, I think I've talked myself out of using the Rust ABI :P

Any thoughts on this before I try to get the rest of the tests using the optimised internals?